refactor auth and route logic
