--- client/components/HomePage/HomePage.js
@@ -1,17 +1,0 @@
-import React, { Component } from 'react';
-
-// components
-import Jumbotron from '../common/jumbotron';
-
-class HomePage extends Component {
-
-  render() {
-    return (
-      <div className="homepage">
-        <Jumbotron />
-      </div>
-    );
-  }
-}
-
-export default HomePage;
