--- client/components/common/Footer.js
@@ -1,19 +1,0 @@
-import React from 'react';
-
-/**
- * Renders the footer element
- * 
- * @returns {object} jsx
- */
-const Footer = () => {
-  return (
-    <footer className="page-footer">
-      <div className="footer-copyright">
-        <div className="container">
-          Copyright © 2017 Andela Socials
-        </div>
-      </div>
-    </footer>
-  );
-}
-export default Footer;
