+++ client/components/common/NotFound.js
@@ -1,13 +1,10 @@
 import React, {Component} from 'react';
 
-class NotFound extends Component {
-  render(){
-    return (
-      <div className="not-found">
-        <p>404</p>
-      </div>
-    );
-  }
-}
+
+const NotFound = () => (
+  <div className="not-found">
+    <p>404</p>
+  </div>
+);
 
 export default NotFound;
