+++ client/public/index.html
@@ -3,11 +3,9 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <meta name="google-signin-client_id" content="1047993778878-tvd9n8v1mag2g4mv1rugkfm016h69ech.apps.googleusercontent.com" />
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" />
     <title> Andela Socials </title>
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css"/>
-    <script src="https://apis.google.com/js/api:client.js"></script>
   </head>
   <body>
     <div id="app"></div>
